Add scheduled pipeline resource (#4)
* Require CIRCLECI_PROJECT to be set for acceptance tests

It's required to run environment variable tests, so bail if unset.

* Update circleci-cli and clean up vendoring

* Add a provider for scheduled pipelines

This works structurally mostly like environment variables, but uses
schedule IDs as internal IDs.

Some hackery has been performed to make organization ID inheritance
from provider settings work. It does work correctly for the most part,
though local state can get a bit confused if the provider setting gets
changed. Explicit organizations on schedules work just fine though.

The scheduled actor ID in this is one of several magic IDs that we
have at CircleCI, and I can guarantee to remain stable.

CRUD operations have been verified to work off the local tree, and
import of existing schedules works as well.

Some provider-side validation is being performed, though it's much
easier to just let the operation fail and print out the API error
message, rather than duplicating all validation we perform in the API
here. An example here is the project<>schedule-name uniqueness
constraint, which is not checked in the provider. Similarly the
requirement for either a branch or tag to be set as part of
parameters, which is actually due to change soon.

* Fix client source

Co-authored-by: Robin Schroer <git@sulami.xyz>